Odometry is now static (+ more)
--This is the first update I have made on GitHub, so all the earlier changes I made on Odometry and everything from the Ultimate Goal Season is not logged here.

The Odometry class, which currently is supposed to use the left and right encoders to track the robot's displacement, is now static. Still figuring out Java, but I think this is the best way to use a class for organization and utility (like the built-in Math class) as opposed to object oriented programming.

Right now, the units are in encoder ticks and I have only implemented the math for two vertical encoders on "dead" wheels. I will try to test this code on a virtual robot and debug it before adding code for three wheels. When this is all complete, I will make documentation explaining how everything works.