Updated all dependencies + Autonomous
Gradle is a thing that automatically downloads files from GitHub that you need to build the app like OpenCV and FTC Dashboard and the latest official FTC robot controller build. But it turns out that I have to actually get the latest versions by manually changing the version numbers. So I did that. Now the map on the FTC dashboard should have the new Freight Frenzy picture and we should get any new changes.
I also moved the new autonomous to a seperate op mode called "GoodAutonomous" which has the shipping element detection in the same file as the autonomous movement.